Update cluster_greyscale_twohead.py
Added int() wrapper in line 123 for dataloader_batch_sz, because when using python3 an integer is expected but the default format of this number is a float.

+++ code/scripts/cluster/cluster_greyscale_twohead.py
@@ -120,7 +120,7 @@ config.twohead = True
 config.in_channels = 1
 config.out_dir = os.path.join(config.out_root, str(config.model_ind))
 assert (config.batch_sz % config.num_dataloaders == 0)
-config.dataloader_batch_sz = config.batch_sz / config.num_dataloaders
+config.dataloader_batch_sz = int(config.batch_sz / config.num_dataloaders)
 
 assert (config.mode == "IID")
 assert ("TwoHead" in config.arch)
